Shop Manager, Scroll Manager, Shop Slot
Finished the basic of Shop Manager, including a refresh method, which is accessed in Scroll Manager to rebuild the scroll layout after the items were added.
Finished the Scroll Manager to manage the scroll properties: rebuild the scroll layout and set value = 1 to the scrollBar so it starts at the start of the item list.
Made the ShopSlot as the same as ItemSlot, but it has two text fields.